requirements.txt
faiss-cpu==1.7.4
langchain==0.0.169
python-dotenv==1.0.0
streamlit==1.22.0